% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addedgemat.R
\name{addedgemat}
\alias{addedgemat}
\title{Edges which can be added to a decomposable graph to give a decomposable graph}
\usage{
addedgemat(M)
}
\arguments{
\item{M}{lower triangular matrix of 1's and 0's, where 1's denote presence
of edges in a decomposable graph}
}
\value{
Out - Lower triangular matrix where 1 denotes edge not present, but can be
added to give a decomposable graph.
}
\description{
For an input M, lower triangular matrix of 1's and 0's, where 1's are
undirected edges of a DECOMPOSABLE graph, addedgemat(M) returns lower
triangular matrix where entry [y,x] is 1 if and only if there is no edge
and adding the edge between y and x returns a decomposable graph.
}
