% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDEdgeSameCCAddable.R
\name{ADDEdgeSameCCAddable}
\alias{ADDEdgeSameCCAddable}
\title{Adding edge in a CC if edge is addable}
\usage{
ADDEdgeSameCCAddable(CG, x, v, w)
}
\arguments{
\item{CG}{an essential graph}

\item{x}{= (x[1],x[2]) two vertices}

\item{v}{the layer in which x[1] and x[2] appear}

\item{w}{the label of the CC in which x[1] and x[2] appear}
}
\value{
Out where
\itemize{
\item Out$CG gives new CG (after addition of edge)
\item Out$Ratio gives ratio of forward to reverse proposals
}
}
\description{
This is called from MCMCmove if x[1],x[2] are in the same
CC, there is no edge and the edge is addable.
All possibilities are considered of an undirected edge, also all
possibilities for
x[1] -> CC2 <- CC1 are also considered where CC1 - CC2 edges exist and
are directed to form immoralities which vanish by legal merger when
x[1] -> CC2 is removed
}
