% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDEdgeSameCCNotAdd.R
\name{ADDEdgeSameCCNotAdd}
\alias{ADDEdgeSameCCNotAdd}
\title{Adding an edge, same CC, if edge is not addable}
\usage{
ADDEdgeSameCCNotAdd(CG, x, v, w)
}
\arguments{
\item{CG}{an essential graph}

\item{x}{= (x[1],x[2]) two vertices}

\item{v}{the layer in which x[1] and x[2] appear}

\item{w}{the label of the CC in which x[1] and x[2] appear}
}
\value{
Out where
\itemize{
\item Out$CG gives new CG (after addition of edge)
\item Out$Ratio gives ratio of forward to reverse proposals
}

export
}
\description{
This is called from MCMCmove if x[1],x[2] are in the same
CC, there is no edge and the edge not addable.
All possibilities are considered of
x[1] -> CC2 <- CC1 where CC1 and CC2 are subsets of CC(w) where
x[2] is in CC2, CC2 connected, CC1 complete set and all CC1 vertices
connected to CC2 vertices, no directed cycles formed when compelled edges
are directed)
}
