% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDSamParDiffCC.R
\name{ADDSamParDiffCC}
\alias{ADDSamParDiffCC}
\title{Add edge, x[1], x[2] from different CCs, but have same parents.}
\usage{
ADDSamParDiffCC(CG, x, v, w)
}
\arguments{
\item{CG}{- an essential graph object}

\item{x}{= (x[1],x[2]) (two vertex labels)}

\item{v}{= (v[1],v[2]) (layers of x[1] resp x[2]) if layers
different; v is a positive integer if both are in same layer}

\item{w}{= (w[1],w[2]) (CC labels for CCs containing x[1] resp. x[2])}
}
\value{
OUT contains a list where
OUT$CG is the new essential graph (after x[1] -> CC2 <- CC1 added
CC1 and CC2 suitable subsets of CC(w[2]) or undirected edge x[1]-x[2]
has been added)
OUT$Ratio is ratio of forward proposal probability divided by
reverse proposal probability
}
\description{
deals with adding edge (directed or undirected) when x[1] and x[2] are from
different CCs with the same parent sets.
For an ordered pair (x[1],x[2]) consider all possibilities of x[1] - x[2]
or x[1] -> CC2 <- CC1
(if undirected edge then *only* to vertex x[2])
}
