% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADDUDdiffCC.R
\name{ADDUDdiffCC}
\alias{ADDUDdiffCC}
\title{Add undirected edge x[1],x[2] different CCs, same parents}
\usage{
ADDUDdiffCC(CG, x, v, w)
}
\arguments{
\item{CG}{- an essential graph object}

\item{x}{= (x[1],x[2]) (two vertex labels)}

\item{v}{is a positive integer giving the layer (same layer
if same parents)}

\item{w}{= (w[1],w[2]) (CC labels for CCs containing x[1] resp. x[2])}
}
\value{
OUT is the new chain graph after the edge has been added (and legal mergers
carried out)
}
\description{
this is called from ADDSamParDiffCC when the
decision is made to add an undirected edge <x[1],x[2]>
x[1] and x[2] in different CCs which have the same parents.
}
