% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADEDiffPar.R
\name{ADEDiffPar}
\alias{ADEDiffPar}
\title{Add directed edge, different CCs and different parent sets}
\usage{
ADEDiffPar(CG, x, v, w)
}
\arguments{
\item{CG}{- an essential graph object}

\item{x}{= (x[1],x[2]) (two vertex labels)}

\item{v}{= (v[1],v[2]) (layers of x[1] resp x[2]) if layers
different; v is a positive integer if both are in same layer}

\item{w}{= (w[1],w[2]) (CC labels for CCs containing x[1] resp. x[2])}
}
\value{
OUT contains a list where
\itemize{
\item OUT$CG is the new essential graph (after x[1] -> S added
where S is a chosen subset of CC(w[2]))
\item OUT$ratio is ratio of forward proposal probability divided by
reverse proposal probability
}
}
\description{
deals with adding directed edge when x[1] and x[2] are from CCs with
different parent sets.
}
