% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCedge.R
\name{CCedge}
\alias{CCedge}
\title{Creates a random tree using Pruefer codes}
\usage{
CCedge(CC)
}
\arguments{
\item{CC}{this is a list of all the chain components}
}
\value{
CCedge list of lower triangular matrices one for each CC
where 1 denotes edge, 0 denotes no edge
}
\description{
This algorithm creates a random tree, each tree equally likely for
each CC. This uses pruefer codes.
}
