% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCsets.R
\name{CCsets}
\alias{CCsets}
\title{List of customers at each table for Chinese Restaurant Process}
\usage{
CCsets(CRP)
}
\arguments{
\item{CRP}{this is the list of table numbers for each vertex (customer)}
}
\value{
L - for each table, the customers at the table.
}
\description{
This algorithm generates the list of customers at each table
of a Chinese Restaurant Process. This is used to generate the chain
components. It is also used to generate the layering for the chain
components
}
