% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCsetsADJ.R
\name{CCsetsADJ}
\alias{CCsetsADJ}
\title{Adjusts Layering to ensure suitable ground layer exists}
\usage{
CCsetsADJ(z, q)
}
\arguments{
\item{z}{collection of CCs}

\item{q}{their layers}
}
\value{
a - the new layering
}
\description{
This algorithm adjusts the layering to ensure that there exists a layer
that is suitable for a ground layer (i.e. so that it is possible to generate
immoralities between layer 1 and layer 2 vertices)
}
