% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCSplit.R
\name{CCSplit}
\alias{CCSplit}
\title{Finds possible splits of a CC for (partial) removal of directed edge}
\usage{
CCSplit(y, Cmat)
}
\arguments{
\item{y}{a vertex which has to be in CC2}

\item{Cmat}{lower triangular, 1's and 0's, 1's denote undirected
edges within the chain component.}
}
\value{
Out where
Out$CC2 lists all possible connected subsets containing y
Out$CC1
Out$CC3
where Out$CC1[[j]], Out$CC2[[j]], Out$CC3[[j]] is jth
possible split.
Out$count gives number of ways.
}
\description{
This is called from RemDE, ADEDiffPar, ADDSamParDiffCC for the purpose of
finding possible connected CC2 (add DE for add moves, remove DE for RemDE)
and a CC1 which is a complete set and all CC2 vertices connected to each
CC1 vertex. (CC3 is simply the remaining vertices not in CC1 or CC2) These
edges become directed CC1 -> CC2 in the add/rem algorithms.
This includes empty set for CC1 - such possibilities have to be removed
for the ADDSamParDiffCC (otherwise directed edges not compelled)
}
