% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCWholeSplit.R
\name{CCWholeSplit}
\alias{CCWholeSplit}
\title{Finds possible splits of a CC for adding within-CC directed edge}
\usage{
CCWholeSplit(x1, x2, Cmat)
}
\arguments{
\item{x1}{a vertex from which new directed edge proceeds.}

\item{x2}{a vertex which has to be in CC2}

\item{Cmat}{lower triangular, 1's and 0's, 1's denote undirected
edges within the chain component.}
}
\value{
Out where
Out$CC2 lists all possibilities for CC2
Out$CC1 lists all possiblilities for CC1
Out$CC3 list of sets of remaining vertices
where Out$CC1[[j]], Out$CC2[[j]], Out$CC3[[j]] is jth
possible split.
Out$count gives number of ways.
}
\description{
This is called from ADDEdgeSameCCNotAdd for the purpose of finding possible
splits of a CC into CC2, CC1, CC3 where CC2 is connected and contains x[2],
CC1 is   a complete set  where all connections
between CC1 and
CC2 exist, CC3 remaining vertices.
}
\details{
CCWholeSplit deals with <x[1],x[2]> not addable, it sends to CCWholeSplitADD
in the case where <x[1],x[2]> is addable.
}
