% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCWholeSplitADD.R
\name{CCWholeSplitADD}
\alias{CCWholeSplitADD}
\title{Finds possible splits of a CC for adding within-CC directed edge}
\usage{
CCWholeSplitADD(x1, x2, Cmat)
}
\arguments{
\item{x1}{a vertex which has to be in CC1}

\item{x2}{a vertex which has to be in CC2}

\item{Cmat}{lower triangular, 1's and 0's, 1's denote undirected
edges within the chain component.}
}
\value{
Out where
Out$CC2 lists all possibilities for CC2
Out$CC1 lists all possiblilities for CC1
Out$CC3 list of sets of remaining vertices
where Out$CC1[[j]], Out$CC2[[j]], Out$CC3[[j]] is jth
possible split.
Out$count gives number of ways.
}
\description{
This is called from ADDEdgeSameCCNotAdd for the purpose of finding possible
splits of a CC into CC2, CC1, CC3 where CC2 is connected and contains x[2],
CC1 is connected and contains a complete set CC11 where all connections
between CC11 and
CC2 exist, CC3 remaining vertices.
}
