% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcomplete.R
\name{checkcomplete}
\alias{checkcomplete}
\title{Checks if a set in the ground layer is complete}
\usage{
checkcomplete(x, CC, ordlay, CCued)
}
\arguments{
\item{x}{- set to be checked}

\item{CC}{- chain components}

\item{ordlay}{- layering of the CCs}

\item{CCued}{- collection of undirected edge matrices for the CCs}
}
\value{
a, which is 1 if complete and 0 otherwise
}
\description{
checks if a set x in layer 1 is complete. This is because, when generating
an essential graph, we use the rejection method (for each vertex in layer 2
we choose a set from layer 1 at random and reject if the set is complete)
}
