% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcompletek.R
\name{checkcompletek}
\alias{checkcompletek}
\title{Checks if a set  in layer k is a complete set.}
\usage{
checkcompletek(x, CC, ordlay, CCued, k)
}
\arguments{
\item{x}{- set to be checked}

\item{CC}{- chain components}

\item{ordlay}{- layering of the CCs}

\item{CCued}{- collection of undirected edge matrices for the CCs}

\item{k}{- layer.}
}
\value{
a = c(1,k,i) if x is complete in layer k, CC labelled i

a = c(0,0,0) if x is not complete
}
\description{
Checks if a set  in layer k is a complete set.
}
