% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkvalid.R
\name{checkvalid}
\alias{checkvalid}
\title{Checks if a candidate set of extra parents is valid}
\usage{
checkvalid(z, i, j, i2ip1, M, CC, CCued, ordlay)
}
\arguments{
\item{z}{is the set we want to check (returned from DEextra)}

\item{i}{is the layer}

\item{j}{number within layer}

\item{i2ip1}{set of previous layer parents (obtained for all CCs)}

\item{M}{candidate 'extra' parents}

\item{CC}{list of all CCs}

\item{CCued}{list of all adjacency matrices for the CCs}

\item{ordlay}{layering of the CCs}
}
\value{
a 1 if valid, 0 if not valid
}
\description{
Checks if a candidate set of extra parents is valid
}
