% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compelled.R
\name{Compelled}
\alias{Compelled}
\title{Directs compelled edges}
\usage{
Compelled(Cmat, MC, PointOut, PointIn)
}
\arguments{
\item{Cmat}{this is the lower triangular directed edge matrix of a CC}

\item{MC}{This is the list of maximal cliques corresponding to Cmat}

\item{PointOut}{This is list of vertices from which arrows are pointing out to
form the new immoralities}

\item{PointIn}{This is vertices into which arrows are pointing}
}
\value{
Out$NewCC this gives a list of the new CCs generated when compelled
edges are appropriately directed

Out$Par This gives a list of the additional parents (due to directing
compelled edges)
}
\description{
After generating new immoralities by adding a directed edge and directing
some undirected edges, this forces direction of compelled edges (previously
undirected) to ensure no other immoralities and no directed or partially
directed cycles
}
