% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete.R
\name{complete}
\alias{complete}
\title{checks if a set of vertices is complete}
\usage{
complete(matr, set)
}
\arguments{
\item{matr}{lower triangular matrix of 0s and 1s indicating undirected
edges of a CC}

\item{set}{a set of vertices within the CC (in CC coordinates)}
}
\value{
ret = 1 if set is a complete set and 0 if it is not a complete set
}
\description{
checks if a set of vertices is complete
}
