% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createpruf.R
\name{createpruf}
\alias{createpruf}
\title{Generates a Pruefer code for each CC}
\usage{
createpruf(CC)
}
\arguments{
\item{CC}{- list of CCs (we only use length of each CC)}
}
\value{
pruf - list of pruefer codes, one for each CC, pruefer code is 0
if CC has length 1 or 2.
}
\description{
takes the CCs and generates a pruefer code for each
returning a single entry of 0 if CC has one or two nodes.
}
