% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRP.R
\name{CRP}
\alias{CRP}
\title{Chinese Restaurant Process}
\usage{
CRP(conc, size)
}
\arguments{
\item{conc}{weight of orange ball; probability of sitting at an empty
table for customer j is conc /(conc + (j-1))}

\item{size}{total number of customers}
}
\value{
x = vector, length = size. x[j] gives table assignment for
customer j.

We use this twice: firstly for CCs, secondly
for layers.

Function is used to generate CCs and to assign CCs to raw layers
}
\description{
assigns customers to tables following a Chinese Restaurant Process
}
