% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEexPr.R
\name{DEexPr}
\alias{DEexPr}
\title{Computes probability of extra parents}
\usage{
DEexPr(CG, p)
}
\arguments{
\item{CG}{- essential graph object}

\item{p}{- 'extra edge' probability}
}
\value{
L - vector of length equal to number of CCs
where L[k] gives the probability for configuration of DEex[[k]] (number of
parents for CC[[k]] that are 2 or more layers back)
}
\description{
Computes probabilities of 'extra edge' configuration (i.e. directed edges
from 2 or more layers back) for each CC.
}
\details{
Routine is called from GraphPR
}
