% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEextra.R
\name{DEextra}
\alias{DEextra}
\title{Generates extra parents}
\usage{
DEextra(i2ip1, CC, CCued, prVIL, ordlay, p)
}
\arguments{
\item{i2ip1}{- directed edges from previous layer (already chosen)}

\item{CC}{- list where CC[[k]] denotes vertices in the kth CC}

\item{CCued}{- list of matrices CCued[[k]] denotes undirected edges
for CC[[k]]}

\item{prVIL}{where prVIL[[k]] denotes the collection of all vertices in
layers 1, ... ,k}

\item{ordlay}{where ordlay [[j]] gives the labels of the CCs in layer j}

\item{p}{- the 'extra edge' probability}
}
\value{
M - list of vectors,
where M[[k]] gives the 2 or more layers back parents for CC[[k]]
}
\description{
This routine generates the 'extra' directed edges - those from two or more
layers back for each CC - ensuring that the i2ip1 edges remain directed
(i.e. not adding edges that will violate Studeny conditions)
}
\details{
If, for a CC, a configuration is chosen that violates Studeny conditions
then we instead keep 'extra edges' empty.

Routine is called from essgraph (generating essential graph)
}
