% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEi2iplus1.R
\name{DEi2iplus1}
\alias{DEi2iplus1}
\title{Generates parents from one layer back}
\usage{
DEi2iplus1(CC, ordlay, VIL, CCued, p)
}
\arguments{
\item{CC}{- list where CC[[k]] denotes vertices in the kth CC}

\item{ordlay}{where ordlay [[j]] gives the labels of the CCs in layer j}

\item{VIL}{where VIL[[k]] denotes the collection of all vertices in
layer k}

\item{CCued}{- list of matrices CCued[[k]] denotes undirected edges
for CC[[k]]}

\item{p}{- the 'extra edge' probability}
}
\value{
L - list of vectors,
where, for CC[[k]] in layer i, M[[k]] gives the parents
from layer i-1

use *rejection method* for layer 1 parents of layer 2 CC (reject if set
of l1 vertices is complete)
}
\description{
This routine generates the directed edges to each CC in layer i from
vertices in layer i-1, for each i.
}
\details{
Routine is called from essgraph (generating essential graph)
}
