% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeposs.R
\name{edgeposs}
\alias{edgeposs}
\title{Indicates if directed edge is possible without creating partially directed cycles}
\usage{
edgeposs(CG, v1, v2, w1, w2)
}
\arguments{
\item{CG}{essential graph object}

\item{v1, }{v2 (layers of x[1] resp x[2])}

\item{w1, }{w2 (CCs containing x[1] resp x[2])}
}
\value{
EDGEYES which is 1 if directed edge x[1] to x[2] is possible
0 otherwise
}
\description{
When x = (x[1],x[2]) is chosen, for x[1] in CC(w1) in layer v1 and
x[2] in CC(w2) in layer v2, this algorithm tells us if a directed
 edge from x[1] to x[2] is possible without creating a directed cycle
 in the CC DAG.
}
