% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/essgraph.R
\name{essgraph}
\alias{essgraph}
\title{Generates random essential graph}
\usage{
essgraph(theta, alpha, p, d)
}
\arguments{
\item{theta}{- Chinese Restaurant concentration parameter for CCs}

\item{alpha}{- Chinese Restaurant concentration parameter for layers}

\item{p}{- extra edge parameter}

\item{d}{- number of vertices}
}
\value{
M
M$CC assignment of vertices to chain components
M$CCued lower triangular matrices, one for each CC
 where 1 denotes undirected edge
M$ordlay list where M$ordlay[[j]] gives the CC labels for layer j
M$VIL where M$VIL[[j]] gives list of vertices in layer j
M$prVIL where M$prVIL[[j]] gives list of vertices from layers 1,...,j
M$DEi2ip1 where  M$DEi2ip1[[j]] gives parents in layer i-1 if CC[[j]]
is in layer i.
M$DEex where M$DEex[[j]] gives parents in layers 1,...,i-2 if CC[[j]]
is in layer i
M$TOTPAR where M$TOTPAR[[j]] gives all parents of CC[[j]]
}
\description{
generates a random essential graph - the undirected edges of each CC
form a spanning tree.
}
