% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal.R
\name{kruskal}
\alias{kruskal}
\title{Runs a Kruskal algorithm}
\usage{
kruskal(C)
}
\arguments{
\item{C}{collection of maximal cliques of a decomposable graph}
}
\value{
J lower triangular matrix 1 if edge in junction tree 0 otherwise.
}
\description{
runs a Kruskal algorithm on a set of maximal cliques of a decomposable graph
to get a junction tree
}
