% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEGMERGE.R
\name{LEGMERGE}
\alias{LEGMERGE}
\title{Legal Merger: performs all possible legal mergers of a quasi-essential graph}
\usage{
LEGMERGE(CG)
}
\arguments{
\item{CG}{a quasi-essential graph object}
}
\value{
NG - the corresponding essential graph.
}
\description{
Start with a quasi-essential graph, perform all possible legal mergers
between CCs to obtain the corresponding essential graph
}
