% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LikScore.R
\name{LikScore}
\alias{LikScore}
\title{Likelihood score for covariance factorised along the graph}
\usage{
LikScore(CG, n, Covarmat)
}
\arguments{
\item{CG}{a chain graph object}

\item{n}{number of instantiations in an n by d data set}

\item{Covarmat}{- d by d empirical covariance matrix for a data set}
}
\value{
LIK log of likelihood score factorising along chain graph,
assuming normality and omitting the contribution from (2 pi) term
}
\description{
Likelihood score for covariance factorised along the graph
}
