% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{MC}
\alias{MC}
\title{Computes the maximal cliques of  decomposable graph}
\usage{
MC(M)
}
\arguments{
\item{M}{the lower triangular matrix defining the undirected decomposable graph.}
}
\value{
MCli  list of maximal cliques
}
\description{
Computes the maximal cliques of  decomposable graph
}
