% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCAdjacent.R
\name{MCAdjacent}
\alias{MCAdjacent}
\title{Adjacency Matrix for Maximal Cliques}
\usage{
MCAdjacent(JTr, W)
}
\arguments{
\item{JTr}{a junction tree computed by Kruskal's algorithm}

\item{W}{the weighted clique graph (edge weights give number of vertices in
intersections)}
}
\value{
Adj - lower triangular matrix 1 if the corresponding
MCs are adjacent in some junction tree and 0 otherwise
}
\description{
This function computes the max clique
adjacency matrix, whereby
two MCs are adjacent if  there is a junction tree in
which they are adjacent
}
