% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPath.R
\name{MCPath}
\alias{MCPath}
\title{Gives a path on the junction between x1 and CC2}
\usage{
MCPath(x1, CC2, C, J)
}
\arguments{
\item{x1}{- a vertex}

\item{CC2}{- a set of vertices}

\item{C}{collection of maximal cliques}

\item{J}{junction tree}
}
\value{
L - path from MC containing x1 to MC which intersects with CC2
}
\description{
Gives a path on the junction between x1 and CC2
}
