% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhd.R
\name{nhd}
\alias{nhd}
\title{Computes the neighbour sets of the vertices}
\usage{
nhd(M)
}
\arguments{
\item{M}{- lower triangular adjacency matrix for a decomposable graph}
}
\value{
O - list
O$L where O$L[[x]] gives neighbours of x, not including x
O$C where O$C[[x]] gives neighbours of x, including x
}
\description{
Input is a lower triangular adjacency matrix for a decomposable graph.
Output gives neighbours.
}
