% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumberComplete.R
\name{NumberComplete}
\alias{NumberComplete}
\title{Computes number of complete sets size k for decomposable graph}
\usage{
NumberComplete(M)
}
\arguments{
\item{M}{adjacency matrix for a decomposable graph}
}
\value{
b a vector, where b[j] gives number of complete sets size j.
}
\description{
input is a matrix M adjacency for decomposable graph
on d vertices. Output is number of complete
sets of size k for k = 1,....,d
}
