% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerfOrd.R
\name{perford}
\alias{perford}
\title{Computes a Perfect Order of the MCs}
\usage{
perford(J)
}
\arguments{
\item{J}{junction tree is output of Kr(C) (Kruskal algorithm)
this routine computes a perfect order of the maximal cliques
for a CONNECTED decomposable graph.}
}
\value{
PO
PO[1,j] gives index of jth MC
PO[2,j] gives index of MC that separates PO[1,j] from earlier MCs
}
\description{
Computes a Perfect Order of the MCs
}
