% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prCC.R
\name{prCC}
\alias{prCC}
\title{Computes probability of undirected edge configuration of the CCs}
\usage{
prCC(CG, p)
}
\arguments{
\item{CG}{an essential graph object
CG$CCs listing of the CCs
CG$CCued listing of lower triangular matrices indicating undirected edges}

\item{p}{extra edge probability parameter}
}
\value{
PC list of the edge configuration probabilities for the CCs
}
\description{
Computes probability of undirected edge configuration of the CCs
}
