% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pri2ip1.R
\name{pri2ip1}
\alias{pri2ip1}
\title{Probability of one-layer-back parent configurations}
\usage{
pri2ip1(CG, p)
}
\arguments{
\item{CG}{an essential graph object}

\item{p}{extra edge parameter}
}
\value{
L list of probabilities of one-layer-back parent
configurations for each CC
}
\description{
computes the probability of the i to i+1 directed edges
}
