% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prlay.R
\name{prlay}
\alias{prlay}
\title{Computes the probability of the layering, given the CCs}
\usage{
prlay(CG, alpha)
}
\arguments{
\item{CG}{an essential graph object}

\item{alpha}{- Chinese Restaurant concentration parameter for layers}
}
\value{
P1 probability of the layering
}
\description{
Computes the probability of the layering, given the CCs
}
