% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progVIL.R
\name{progVIL}
\alias{progVIL}
\title{computes union of layer together with all previous layers}
\usage{
progVIL(VIL)
}
\arguments{
\item{VIL}{- list where VIL[[j]] gives vertices in each layer}
}
\value{
g3 where g3[[j]] is union of VIL[[i]] for i=1,...,j
}
\description{
computes union of layer together with all previous layers
}
