% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prs.R
\name{prs}
\alias{prs}
\title{Computes probability of table assignments according to Chinese Restaurant Process}
\usage{
prs(s, theta)
}
\arguments{
\item{s:}{vector of lengths of the tables}

\item{theta:}{weight of 'orange ball' for Polya urn}
}
\value{
pr: probability of the table assignments according to
Ewens sampling formula with parameter theta.
}
\description{
Computes probability of table assignments according to Chinese Restaurant Process
}
