% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruefer2edge.R
\name{pruefer2edge}
\alias{pruefer2edge}
\title{Computes the lower triangular adjacency matrix for tree defined by Pruefer code.}
\usage{
pruefer2edge(p2e, CComp)
}
\arguments{
\item{p2e:}{a pruefer code of length n-2}

\item{CComp:}{corresponding chain component of length n}
}
\value{
mat: an n by n lower triangular adjacency matrix where
1's correspond to edges given by the pruefer code.
}
\description{
Computes the lower triangular adjacency matrix for tree defined by Pruefer code.
}
