% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemDE.R
\name{RemDE}
\alias{RemDE}
\title{Removes directed edge}
\usage{
RemDE(CG, x, v, w)
}
\arguments{
\item{CG}{- an essential graph object}

\item{x}{= (x[1],x[2]) (two vertex labels)}

\item{v}{= (v[1],v[2]) (layers of x[1] resp x[2]) if layers
different; v is a positive integer if both are in same layer}

\item{w}{= (w[1],w[2]) (CC labels for CCs containing x[1] resp. x[2])}
}
\value{
OUT contains a list where
OUT$CG is the new essential graph (after x[1] -> S removed
S suitable subset of CC(w[2]))
OUT$ratio is ratio of forward proposal probability divided by
reverse proposal probability
}
\description{
deals with removing a directed edge from x[1] to S, subset CC(w[2]).
We find all splits of CC(w[2]) into disjoint subsets CC2, CC1, CC3 such
that their union is CC(w[2]), x[2] in CC2, CC2 connected,
CC1 complete, all connections between CC1 and CC2 present. Then we direct
CC1 -> CC2
}
\details{
Adding edge x[1] to CC2 reverses the move.
}
