% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMUE.R
\name{RMUE}
\alias{RMUE}
\title{Removes undirected edge}
\usage{
RMUE(CG, x, v, w)
}
\arguments{
\item{CG}{- an essential graph object}

\item{x}{= (x[1],x[2]) two vertices in the same CC}

\item{v}{- the layer}

\item{w}{- the CC}
}
\value{
: Out
Out$CG is the new essential graph
Out$Ratio is the ratio between reverse and forward proposals
}
\description{
Removes undirected edge
}
