% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMUESameCC.R
\name{RMUESameCC}
\alias{RMUESameCC}
\title{Removes undirected edge where removal does not split the CC into two CCs}
\usage{
RMUESameCC(CG, x, v, w)
}
\arguments{
\item{CG}{- an essential graph object}

\item{x}{= (x[1],x[2]) two vertices in the same CC}

\item{v}{- the layer}

\item{w}{- the CC}
}
\value{
: Out where
\itemize{
\item Out$CG is the new essential graph
\item Out$Ratio is the ratio between reverse and forward proposals
}
}
\description{
Removes undirected edge where removal does not split the CC into two CCs
}
