% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler.R
\name{Sampler}
\alias{Sampler}
\title{Metropolis sampler for essential graphs from prior distribution}
\usage{
Sampler(N, d, theta, alpha, p)
}
\arguments{
\item{N}{is the number of MCMC iterations}

\item{d}{is the number of vertices}

\item{theta}{is the assign vertices to CCs parameter for CRP}

\item{alpha}{is the layering CCs parameter for Chinese Restaurant Process}

\item{p}{is the edge probability for the prior}
}
\value{
OUT contains a list of N, each with the following components
\itemize{
\item OUT[[j]]$OldCG is the input CG from the previous move
\item OUT[[j]]$NewCG is the important one - the new CG.
  which is the proposal if the proposal is accepted
  or the OldCG if the proposal is rejected
\item OUT[[j]]$RatioProp is   P(NewCG -> OldCG)/P(OldCG -> NewCG)
  which is the ratio of probabilities of proposing old from new to new
  from old
\item OUT[[j]]$RatioGraphs is the ratio of the prior probabilities of new to old graphs
\item OUT[[j]]$Score is the logaritm of the (prior) probability of NewCG
}
}
\description{
This algorithm simply runs a Metropolis Hastings sampler through essential
graphs using 'prior' for 'posterior' and thus may be used for sampling
according to the given 'prior'.
}
