% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCovariance.R
\name{SamplerCovariance}
\alias{SamplerCovariance}
\title{Metropolis  Sampler: input is covariance}
\usage{
SamplerCovariance(covar, n, N, theta, alpha, p)
}
\arguments{
\item{covar}{is a d by d covariance matrix with diagonal elements all 1.}

\item{n}{is the number of observations (for which covariance is likelihood raised
to n/2)}

\item{N}{is the number of MCMC iterations}

\item{theta}{is the assign vertices to CCs parameter for CRP}

\item{alpha}{is the layering CCs parameter for Chinese Restaurant Process}

\item{p}{is the edge probability parameter}
}
\value{
OUT contains a list of N, each with the following components
\itemize{
\item OUT[[j]]$OldCG current graph (before proposal)
\item OUT[[j]]$NewCG new graph
\item OUT[[j]]$logLikelihood log of the likelihood score (missing 2 pi)
\item OUT[[j]]$GraphScore log of probability of new graph
\item OUT[[j]]$WholeScore sum of GraphScore and logLikelihood
\item OUT[[j]]$move 1 if proposal accepted, 0 otherwise
}
}
\description{
Metropolis  Sampler: input is covariance
}
