% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerData.R
\name{SamplerData}
\alias{SamplerData}
\title{Metropolis  sampler: input is data matrix}
\usage{
SamplerData(data, N, theta, alpha, p)
}
\arguments{
\item{data}{is an n by d data matrix n instantiations of d-variate vector,
which we centre and scale. We wrote our own (simple) centre/scale routine
because the built-in routine seemed to give the wrong answer.}

\item{N}{is the number of MCMC iterations}

\item{theta}{is the assign vertices to CCs parameter for CRP}

\item{alpha}{is the layering CCs parameter for Chinese Restaurant Process}

\item{p}{is the edge probability parameter}
}
\value{
OUT contains a list of N, each with the following components
OUT$OldCG is the input essential graph
OUT$CG is the essential graph (CG = chain graph)
OUT$Post is the posterior score for the graph
OUT$Prior is the prior graph probability
OUT$Lik is the likelihood score
}
\description{
Metropolis  sampler: input is data matrix
}
