% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Separators.R
\name{Separators}
\alias{Separators}
\title{Computes list of separators}
\usage{
Separators(M)
}
\arguments{
\item{M}{is the lower triangular adjacency matrix for a connected
decomposable graph}
}
\value{
S the list of separators (according to multiplicity)
}
\description{
Computes list of separators
}
