% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertinlayer.R
\name{vertinlayer}
\alias{vertinlayer}
\title{Union of all CCs in each layer}
\usage{
vertinlayer(ordlay, CC)
}
\arguments{
\item{ordlay}{- which CCs are in each layer}

\item{CC}{- collection of chain components}
}
\value{
g2 = list g2[[i]] all vertices in layer i
}
\description{
This simply takes the union of all CCs in a layer, for each layer.
}
