% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightedCliqueGraph.R
\name{WeightedCliqueGraph}
\alias{WeightedCliqueGraph}
\title{Computes the weighted maximal clique graph}
\usage{
WeightedCliqueGraph(C)
}
\arguments{
\item{C}{collection of maximal cliques of a decomposable graph}
}
\value{
W lower triangular graph W[i,j] = |C[[i]] intersect C[[j]]|
for i > j
}
\description{
This function takes a collection of maximal cliques
and returns a lower triangular matrix W
where W[i,j] for i > j is the number of vertices in
C[[i]] intersect C[[j]]
}
