% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphPR.R
\name{GraphPR}
\alias{GraphPR}
\title{Computes the probability of an essential graph}
\usage{
GraphPR(GRAPH, theta, alpha, p)
}
\arguments{
\item{GRAPH}{an essential graph object}

\item{theta}{- Chinese Restaurant concentration parameter for CCs}

\item{alpha}{- Chinese Restaurant concentration parameter for layers}

\item{p}{- extra edge parameter}
}
\value{
L
L$PCCs probability of assignment of vertices to CCs
L$PCCued vector, length is number of CCs, each entry is probability
of edge configuration for the CC
L$Pordlay probability of assignment of CCs to layers (conditioned on CCs)
L$PDEi2ip1 vector, each entry is probability of one-layer-back parent sets
L$PDEex vector, each entry is probability of two-or-more layer back
parent sets (conditioned on preceding probabilities)

All the numbers, when multiplied together, give the probability
of the essential graph.
}
\description{
Computes the probability of an essential graph
}
