% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCmove.R
\name{MCMCmove}
\alias{MCMCmove}
\title{Randomly selects two different vertices and computes a proposal.}
\usage{
MCMCmove(CG)
}
\arguments{
\item{CG}{an essential graph object}
}
\value{
Out  a list where
Out$OldCG = CG (the original - input - essential graph)
Out$NewCG <- NewCG (the proposed graph)
Out$Ratio <- Ratio (ratio of reverse proposal to forward proposal)
}
\description{
Randomly selects two different vertices and computes a proposal.
}
