% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sampler.R
\name{Sampler}
\alias{Sampler}
\title{Metropolis sampler for essential graphs from prior distribution}
\usage{
Sampler(N, d, theta, alpha, p)
}
\arguments{
\item{N}{is the number of MCMC iterations}

\item{d}{is the number of vertices}

\item{theta}{is the assign vertices to CCs parameter for CRP}

\item{alpha}{is the layering CCs parameter for Chinese Restaurant Process}

\item{p}{is the edge probability for the prior}
}
\value{
OUT contains a list of N, each with three components
OUT$OldCG is the input CG from the previous move
OUT$NewCG is the important one - the new CG.
  This is the proposal if the proposal is accepted
  or the OldCG if the proposal is rejected
OUT$RatioProp is the ratio P(NewCG -> OldCG)/P(OldCG -> NewCG)
  numerator probability of proposing OldCG from new
  denominator probability of proposing NewCG from old
OUT$RatioGraphs is the ratio of the prior probabilities of new to old graphs
OUT$Score is the prior probability of NewCG
}
\description{
This algorithm simply runs a Metropolis Hastings sampler through essential
graphs using 'prior' for 'posterior' and thus may be used for sampling
according to the given 'prior'.
}
