% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerCovariance.R
\name{SamplerCovariance}
\alias{SamplerCovariance}
\title{Metropolis  Sampler: input is covariance}
\usage{
SamplerCovariance(covar, n, N, theta, alpha, p)
}
\arguments{
\item{covar}{is a d by d covariance matrix with diagonal elements all 1.}

\item{n}{is the number of observations (for which covariance is likelihood raised
to n/2)}

\item{N}{is the number of MCMC iterations}

\item{theta}{is the assign vertices to CCs parameter for CRP}

\item{alpha}{is the layering CCs parameter for Chinese Restaurant Process}

\item{p}{is the edge probability parameter}
}
\value{
OUT contains a list of N, each with three components
OUT$CG is the essential graph (CG = chain graph)
OUT$Post is the posterior score for the graph
OUT$Prior is the prior graph probability
OUT$Lik is the likelihood score
}
\description{
Metropolis  Sampler: input is covariance
}
