% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplerData.R
\name{SamplerData}
\alias{SamplerData}
\title{Metropolis  sampler: input is data matrix}
\usage{
SamplerData(data, N, theta, alpha, p)
}
\arguments{
\item{data}{is an n by d data matrix n instantiations of d-variate vector,
which we centre and scale. We wrote our own (simple) centre/scale routine
because the built-in routine seemed to give the wrong answer.}

\item{N}{is the number of MCMC iterations}

\item{theta}{is the assign vertices to CCs parameter for CRP}

\item{alpha}{is the layering CCs parameter for Chinese Restaurant Process}

\item{p}{is the edge probability parameter}
}
\value{
OUT a list of length N, where OUT[[j]] has following components:
\itemize{
\item OUT[[j]]$OldCG current graph (before proposal)
\item OUT[[j]]$NewCG new graph
\item OUT[[j]]$logLikelihood log of the likelihood score (missing 2 pi)
\item OUT[[j]]$GraphScore log of probability of new graph
\item OUT[[j]]$WholeScore sum of GraphScore and logLikelihood
\item OUT[[j]]$move 1 if proposal accepted, 0 otherwise
}
}
\description{
Metropolis  sampler: input is data matrix
}
